支持密码

+++ run.sh
@@ -3,28 +3,40 @@
 IPV4="203.0.113.0"
 IPV6="2001:db8:ffff:ffff:ffff:ffff:ffff:ffff"
 
+# 造一个错误命令来进入循环
+curl 333.333.333.333 2>/dev/null
+while [ $? -ne 0 ] 
+do
+    PASSHASH=$(htpasswd -bnBC 10 "" $PASSWORD | tr -d ':\n' | sed 's/$2y/$2a/')
+    sed -i "s/\$2a\$10\$UL51kOUMX6uwmI1y8ddCje.5IQ0.FM4aeTbjC5n3N8Vu9QFZnX.Qq/${PASSHASH}/g" /opt/AdGuardHome/AdGuardHome.yaml
+done
+
 while [[ true ]]; do
     sleep 1
 
     C=$(curl -4 ifconfig.co 2>/dev/null)
-    if [[ "$IPV4" != "$C" ]];then
-        sed -i "s/$IPV4/$C/g" /opt/AdGuardHome/AdGuardHome.yaml
-        killall -9 AdGuardHome >/dev/null 2>&1 &
-        killall -9 sniproxy >/dev/null 2>&1 &
-        sleep 1
-        /opt/AdGuardHome/AdGuardHome --no-check-update -c /opt/AdGuardHome/AdGuardHome.yaml 2>&1 &
-        /usr/sbin/sniproxy -c /etc/sniproxy.conf -f 2>&1 &
-        IPV4=$C
+    if [ $? -eq 0 ];then
+        if [[ "$IPV4" != "$C" ]];then
+            sed -i "s/$IPV4/$C/g" /opt/AdGuardHome/AdGuardHome.yaml
+            killall -9 AdGuardHome >/dev/null 2>&1 &
+            killall -9 sniproxy >/dev/null 2>&1 &
+            sleep 1
+            /opt/AdGuardHome/AdGuardHome --no-check-update -c /opt/AdGuardHome/AdGuardHome.yaml 2>&1 &
+            /usr/sbin/sniproxy -c /etc/sniproxy.conf -f 2>&1 &
+            IPV4=$C
+        fi
     fi
 
     C=$(curl -6 ifconfig.co 2>/dev/null)
-    if [[ "$IPV6" != "$C" ]];then
-        sed -i "s/$IPV6/$C/g" /opt/AdGuardHome/AdGuardHome.yaml
-        killall -9 AdGuardHome >/dev/null 2>&1 &
-        killall -9 sniproxy >/dev/null 2>&1 &
-        sleep 1
-        /opt/AdGuardHome/AdGuardHome --no-check-update -c /opt/AdGuardHome/AdGuardHome.yaml 2>&1 &
-        /usr/sbin/sniproxy -c /etc/sniproxy.conf -f 2>&1 &
-        IPV6=$C
+    if [ $? -eq 0 ];then
+        if [[ "$IPV6" != "$C" ]];then
+            sed -i "s/$IPV6/$C/g" /opt/AdGuardHome/AdGuardHome.yaml
+            killall -9 AdGuardHome >/dev/null 2>&1 &
+            killall -9 sniproxy >/dev/null 2>&1 &
+            sleep 1
+            /opt/AdGuardHome/AdGuardHome --no-check-update -c /opt/AdGuardHome/AdGuardHome.yaml 2>&1 &
+            /usr/sbin/sniproxy -c /etc/sniproxy.conf -f 2>&1 &
+            IPV6=$C
+        fi
     fi
 done
